Improve the code structure of RandomTest.java, and test bigger number of arguments.
