+++ test/RandomTest.java
@@ -32,10 +32,10 @@ public class RandomTest {
     private static final long ENERGY_LIMIT = 10_000_000L;
     private static final long ENERGY_PRICE = 1L;
 
+
     @Before
     public void setup() {
         compiler = new ABICompiler();
-        setupRandomArgumentsEnv();
     }
 
     private Address installTestDApp() {
@@ -76,59 +76,63 @@ public class RandomTest {
     }
 
     @Test
-    public void testFixedArguments() {
+    public void testArgumentsWithFixedTypesRandomValues() {
+        RandomArgumentsGenerator argsGenerator = new RandomArgumentsGenerator();
+        argsGenerator.addFixedTypesAndRandomValues(18);
+
         String className = "RandomClass";
-        String methodName = getRandomString(9);
-        methodName = "test";
 
-        for (int i = 0; i < numberOfArgs; i++) {
-            addRandomType(i+1, i);
-        }
+        String methodName = argsGenerator.getRandomString(9);
 
-        byte[] jar = JarBuilder.buildJarForExplicitClassNamesAndBytecode(className, createClass(methodName), new HashMap<>());
+        System.out.println(String.format("%s.%s%s --- %s arguments", className , methodName,
+                        RandomTest.getMethodDescriptor(String.join("", argsGenerator.getArgDescriptors())), argsGenerator.getNumberOfArgs()));
+
+        byte[] jar = JarBuilder.buildJarForExplicitClassNamesAndBytecode(className, createClass(methodName, argsGenerator), new HashMap<>());
         compiler.compile(new ByteArrayInputStream(jar));
 
         Address dapp = installTestDApp();
 
-        byte[] result = (byte[]) callStatic(dapp, methodName, this.varArgs);
-        byte[] expected = ABIEncoder.encodeMethodArguments("",this.varArgs);
+        byte[] result = (byte[]) callStatic(dapp, methodName, argsGenerator.getVarArgs());
+        byte[] expected = ABIEncoder.encodeMethodArguments("",argsGenerator.getVarArgs());
         Assert.assertTrue(Arrays.equals(expected, result));
     }
 
     @Test
-    public void testRandoms() {
-        int times = 10;
+    public void testArgumentsWithRandomTypesRandomValue() {
+        int times = 100;
         for(int i = 0; i < times; ++i) {
             testRandom();
         }
     }
 
     private void testRandom() {
+        RandomArgumentsGenerator argsGenerator = new RandomArgumentsGenerator();
+        argsGenerator.addRandomTypeAndRandomValues(50);
+
         String className = "RandomClass";
-        String methodName = getRandomString(9);
-        methodName = "test";
 
-        for (int i = 0; i < numberOfArgs; i++) {
-            addRandomType(random.nextInt(18) + 1, i);
-        }
+        String methodName = argsGenerator.getRandomString(9);
 
-        byte[] jar = JarBuilder.buildJarForExplicitClassNamesAndBytecode(className, createClass(methodName), new HashMap<>());
+        System.out.println(String.format("%s.%s%s --- %s arguments", className , methodName,
+                RandomTest.getMethodDescriptor(String.join("", argsGenerator.getArgDescriptors())), argsGenerator.getNumberOfArgs()));
+
+        byte[] jar = JarBuilder.buildJarForExplicitClassNamesAndBytecode(className, createClass(methodName, argsGenerator), new HashMap<>());
         compiler.compile(new ByteArrayInputStream(jar));
 
         Address dapp = installTestDApp();
 
-        byte[] result = (byte[]) callStatic(dapp, methodName, this.varArgs);
-        byte[] expected = ABIEncoder.encodeMethodArguments("",this.varArgs);
+        byte[] result = (byte[]) callStatic(dapp, methodName, argsGenerator.getVarArgs());
+        byte[] expected = ABIEncoder.encodeMethodArguments("",argsGenerator.getVarArgs());
         Assert.assertTrue(Arrays.equals(expected, result));
     }
 
-    private byte[] createClass(String methodName) {
+    private byte[] createClass(String methodName, RandomArgumentsGenerator argsGenerator) {
         ClassWriter classWriter = new ClassWriter(0);
         classWriter.visit(V10, ACC_PUBLIC + ACC_SUPER, "RandomClass", null, "java/lang/Object", null);
 
         MethodVisitor methodVisitor = classWriter.visitMethod(ACC_PUBLIC | ACC_STATIC,
                 methodName,
-                RandomTest.getMethodDescriptor(String.join("", this.argDescriptors)),
+                RandomTest.getMethodDescriptor(String.join("", argsGenerator.getArgDescriptors())),
                 null,
                 null);
         AnnotationVisitor annotationVisitor = methodVisitor.visitAnnotation("Lorg/aion/abigenerator/Callable;", false);
@@ -139,15 +143,15 @@ public class RandomTest {
         methodVisitor.visitLabel(label0);
         methodVisitor.visitLdcInsn("");
 
-        methodVisitor.visitIntInsn(BIPUSH, this.argTypes.length);
+        methodVisitor.visitIntInsn(BIPUSH, argsGenerator.getArgTypes().length);
         methodVisitor.visitTypeInsn(ANEWARRAY, "java/lang/Object");
         int index = 0;
-        for(int i = 0; i < this.argDescriptors.length; ++i, ++index) {
+        for(int i = 0; i < argsGenerator.getArgDescriptors().length; ++i, ++index) {
             methodVisitor.visitInsn(DUP);
             methodVisitor.visitIntInsn(BIPUSH, i);
-            loadLocalVariableToStack(methodVisitor, this.argTypes[i], index);
-            if(this.argTypes[i] == 6 || this.argTypes[i] == 8) index++;
-            castArgumentType(methodVisitor, this.argTypes[i]);
+            loadLocalVariableToStack(methodVisitor, argsGenerator.getArgTypes()[i], index);
+            if(argsGenerator.getArgTypes()[i] == Type.LONG || argsGenerator.getArgTypes()[i] == Type.DOUBLE) index++;
+            castArgumentType(methodVisitor, argsGenerator.getArgTypes()[i]);
             methodVisitor.visitInsn(AASTORE);
         }
 
@@ -158,10 +162,10 @@ public class RandomTest {
         methodVisitor.visitLabel(label1);
 
         index = 0;
-        for(int i = 0; i < this.argDescriptors.length; ++i, ++index) {
-            System.out.println(String.format("%d %s %s", this.argTypes[i], getTypeDescriptor(this.argTypes[i]), this.varArgs[i].toString()));
-            methodVisitor.visitLocalVariable("v" + String.valueOf(i+1), getTypeDescriptor(this.argTypes[i]), null, label0, label1, index);
-            if(this.argTypes[i] == 6 || this.argTypes[i] == 8) index++;
+        for(int i = 0; i < argsGenerator.getArgDescriptors().length; ++i, ++index) {
+            System.out.println(String.format("%s %s %s", argsGenerator.getArgTypes()[i], getTypeDescriptor(argsGenerator.getArgTypes()[i]), argsGenerator.getVarArgs()[i]));
+            methodVisitor.visitLocalVariable("v" + String.valueOf(i+1), getTypeDescriptor(argsGenerator.getArgTypes()[i]), null, label0, label1, index);
+            if(argsGenerator.getArgTypes()[i] == Type.LONG || argsGenerator.getArgTypes()[i] == Type.DOUBLE) index++;
         }
 
         methodVisitor.visitMaxs(index, index);
@@ -180,25 +184,22 @@ public class RandomTest {
         return classWriter.toByteArray();
     }
 
-    private void loadLocalVariableToStack(MethodVisitor methodVisitor, int type, int index) {
+    private void loadLocalVariableToStack(MethodVisitor methodVisitor, Type type, int index) {
         switch (type) {
-            case 1: //byte
-            case 2: //boolean
-            case 3: //char
-            case 4: //short
-            case 5: //int
+            case BYTE:
+            case BOOLEAN:
+            case CHAR:
+            case SHORT:
+            case INT:
                 methodVisitor.visitVarInsn(ILOAD, index);
                 break;
-            case 6:
-                //long
+            case LONG:
                 methodVisitor.visitVarInsn(LLOAD, index);
                 break;
-            case 7:
-                //float
+            case FLOAT:
                 methodVisitor.visitVarInsn(FLOAD, index);
                 break;
-            case 8:
-                //double
+            case DOUBLE:
                 methodVisitor.visitVarInsn(DLOAD, index);
                 break;
             default:
@@ -208,41 +209,34 @@ public class RandomTest {
         }
     }
 
-    private void castArgumentType(MethodVisitor methodVisitor, int type) {
+    private void castArgumentType(MethodVisitor methodVisitor, Type type) {
         switch (type) {
-            case 1:
-                //byte
+            case BYTE:
                 methodVisitor.visitMethodInsn(INVOKESTATIC, "java/lang/Byte", "valueOf", "(B)Ljava/lang/Byte;", false);
                 break;
-            case 2:
-                //boolean
+            case BOOLEAN:
                 methodVisitor.visitMethodInsn(INVOKESTATIC, "java/lang/Boolean", "valueOf", "(Z)Ljava/lang/Boolean;", false);
                 break;
-            case 3:
-                //char
+            case CHAR:
                 methodVisitor.visitMethodInsn(INVOKESTATIC, "java/lang/Character", "valueOf", "(C)Ljava/lang/Character;", false);
                 break;
-            case 4:
-                //short
+            case SHORT:
                 methodVisitor.visitMethodInsn(INVOKESTATIC, "java/lang/Short", "valueOf", "(S)Ljava/lang/Short;", false);
                 break;
-            case 5:
-                //int
+            case INT:
                 methodVisitor.visitMethodInsn(INVOKESTATIC, "java/lang/Integer", "valueOf", "(I)Ljava/lang/Integer;", false);
                 break;
-            case 6:
-                //long
+            case LONG:
                 methodVisitor.visitMethodInsn(INVOKESTATIC, "java/lang/Long", "valueOf", "(J)Ljava/lang/Long;", false);
                 break;
-            case 7:
-                //float
+            case FLOAT:
                 methodVisitor.visitMethodInsn(INVOKESTATIC, "java/lang/Float", "valueOf", "(F)Ljava/lang/Float;", false);
                 break;
-            case 8:
-                //double
+            case DOUBLE:
                 methodVisitor.visitMethodInsn(INVOKESTATIC, "java/lang/Double", "valueOf", "(D)Ljava/lang/Double;", false);
                 break;
             default:
+                //other
                 break;
         }
     }
@@ -251,373 +245,64 @@ public class RandomTest {
         return String.format("(%s)[B", argumentsDescriptor);
     }
 
-    private String getRandomArgumentsDescriptor(int argumentsNumber) {
-
-        Random random = new Random();
-        StringBuilder stringBuilder = new StringBuilder();
-        stringBuilder.append("\"(");
-        for (int i = 0; i < argumentsNumber; ++i) {
-            stringBuilder.append(getTypeDescriptor(random.nextInt(18)));
-        }
-        stringBuilder.append(")Z");
-        System.out.println(stringBuilder.toString());
-        return stringBuilder.toString();
-    }
-
-    private String getTypeDescriptor(int type) {
+    private String getTypeDescriptor(Type type) {
         String t = "";
         switch (type) {
-            case 1:
-                //byte
+            case BYTE:
                 t = "B";
                 break;
-            case 2:
-                //bool
+            case BOOLEAN:
                 t = "Z";
                 break;
-            case 3:
-                //char
+            case CHAR:
                 t = "C";
                 break;
-            case 4:
-                //short
+            case SHORT:
                 t = "S";
                 break;
-            case 5:
-                //int
+            case INT:
                 t = "I";
                 break;
-            case 6:
-                //long
+            case LONG:
                 t = "J";
                 break;
-            case 7:
-                //float
+            case FLOAT:
                 t = "F";
                 break;
-            case 8:
-                //double
+            case DOUBLE:
                 t = "D";
                 break;
-            case 9:
-                //byte array
+            case BYTE_ARRAY:
                 t = "[B";
                 break;
-            case 10:
-                //bool array
+            case BOOL_ARRAY:
                 t = "[Z";
                 break;
-            case 11:
-                //char array
+            case CHAR_ARRAY:
                 t = "[C";
                 break;
-            case 12:
-                //short array
+            case SHORT_ARRAY:
                 t = "[S";
                 break;
-            case 13:
-                //int array
+            case INT_ARRAY:
                 t = "[I";
                 break;
-            case 14:
-                //long array
+            case LONG_ARRAY:
                 t = "[J";
                 break;
-            case 15:
-                //float array
+            case FLOAT_ARRAY:
                 t = "[F";
                 break;
-            case 16:
-                //double array
+            case DOUBLE_ARRAY:
                 t = "[D";
                 break;
-            case 17:
-                //string
+            case STRING:
                 t = "Ljava/lang/String;";
                 break;
-            case 18:
-                //int array 2d
+            case INT_ARRAY_2D:
                 t = "[[I";
                 break;
         }
         return t;
-    }
-
-    private static Random random;
-
-    private int numberOfArgs;
-    private Object[] varArgs;
-    private ABICodec.Tuple[] argTuples;
-    private int[] argTypes;
-    private String[] argDescriptors;
-    private static final int upperBoundOfNumOfArgs = 10;
-
-    public void setupRandomArgumentsEnv() {
-        random = new Random();
-        long seed = random.nextLong();
-        System.out.println("Test seed is " + seed);
-        random.setSeed(seed);
-
-        numberOfArgs = random.nextInt(RandomTest.upperBoundOfNumOfArgs);
-        varArgs = new Object[numberOfArgs];
-        argTuples = new ABICodec.Tuple[numberOfArgs];
-        argTypes = new int[numberOfArgs];
-        argDescriptors = new String[numberOfArgs];
-    }
-
-    private static String getRandomString(int max) {
-        int length = random.nextInt(max) + 5;
-        String allChars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz1234567890";
-
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < length; i++) {
-            sb.append(allChars.charAt(random.nextInt(62)));
-        }
-        return sb.toString();
-    }
-
-    private void addByte(int i) {
-        byte b = (byte) random.nextInt();
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(Byte.class, b);
-    }
-
-    private void addBool(int i) {
-        boolean b = random.nextBoolean();
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(Boolean.class, b);
-    }
-
-    private void addChar(int i) {
-        char b = (char) random.nextInt();
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(Character.class, b);
-    }
-
-    private void addShort(int i) {
-        short b = (short) random.nextInt();
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(Short.class, b);
-    }
-
-    private void addInt(int i) {
-        int b = random.nextInt();
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(Integer.class, b);
-    }
-
-    private void addLong(int i) {
-        long b = random.nextLong();
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(Long.class, b);
-    }
-
-    private void addFloat(int i) {
-        float b = random.nextFloat();
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(Float.class, b);
-    }
-
-    private void addDouble(int i) {
-        double b = random.nextDouble();
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(Double.class, b);
-    }
-
-    private void addByteArray(int i) {
-        byte[] b = new byte[random.nextInt(50)];
-        random.nextBytes(b);
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(byte[].class, b);
-    }
-
-    private void addBoolArray(int i) {
-        int len = random.nextInt(50);
-        boolean[] b = new boolean[len];
-        for (int j = 0; j < len; j++) {
-            b[j] = random.nextBoolean();
-        }
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(boolean[].class, b);
-    }
-
-    private void addCharArray(int i) {
-        int len = random.nextInt(50);
-        char[] b = new char[len];
-        for (int j = 0; j < len; j++) {
-            b[j] = (char) random.nextInt();
-        }
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(char[].class, b);
-    }
-
-    private void addShortArray(int i) {
-        int len = random.nextInt(50);
-        short[] b = new short[len];
-        for (int j = 0; j < len; j++) {
-            b[j] = (short) random.nextInt();
-        }
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(short[].class, b);
-    }
-
-    private void addIntArray(int i) {
-        int[] b = getRandomIntArray();
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(int[].class, b);
-    }
-
-    private void addLongArray(int i) {
-        int len = random.nextInt(50);
-        long[] b = new long[len];
-        for (int j = 0; j < len; j++) {
-            b[j] = random.nextLong();
-        }
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(long[].class, b);
-    }
-
-    private void addFloatArray(int i) {
-        int len = random.nextInt(50);
-        float[] b = new float[len];
-        for (int j = 0; j < len; j++) {
-            b[j] = random.nextFloat();
-        }
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(float[].class, b);
-    }
-
-    private void addDoubleArray(int i) {
-        int len = random.nextInt(50);
-        double[] b = new double[len];
-        for (int j = 0; j < len; j++) {
-            b[j] = random.nextDouble();
-        }
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(double[].class, b);
-    }
-
-    private void addString(int i) {
-        String b = getRandomString(20);
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(String.class, b);
-    }
-
-    private int[] getRandomIntArray() {
-        int len = random.nextInt(50);
-        int[] b = new int[len];
-        for (int j = 0; j < len; j++) {
-            if (random.nextInt(10) < 8) {
-                b[j] = random.nextInt();
-            }
-        }
-        return b;
-    }
-
-    private void addIntArray2D(int i) {
-        int len = random.nextInt(50);
-        int[][] b = new int[len][];
-
-        for (int j = 0; j < len; j++) {
-            b[j] = getRandomIntArray();
-        }
-
-        varArgs[i] = b;
-        argTuples[i] = new ABICodec.Tuple(int[][].class, b);
-    }
-
-    private void addRandomType(int type, int index) {
-        switch (type) {
-            case 1:
-                addByte(index);
-                break;
-            case 2:
-                addBool(index);
-                break;
-            case 3:
-                addChar(index);
-                break;
-            case 4:
-                addShort(index);
-                break;
-            case 5:
-                addInt(index);
-                break;
-            case 6:
-                addLong(index);
-                break;
-            case 7:
-                addFloat(index);
-                break;
-            case 8:
-                addDouble(index);
-                break;
-            case 9:
-                addByteArray(index);
-                break;
-            case 10:
-                addBoolArray(index);
-                break;
-            case 11:
-                addCharArray(index);
-                break;
-            case 12:
-                addShortArray(index);
-                break;
-            case 13:
-                addIntArray(index);
-                break;
-            case 14:
-                addLongArray(index);
-                break;
-            case 15:
-                addFloatArray(index);
-                break;
-            case 16:
-                addDoubleArray(index);
-                break;
-            case 17:
-                addString(index);
-                break;
-            case 18:
-                addIntArray2D(index);
-                break;
-        }
-        argTypes[index] = type;
-        argDescriptors[index] = getTypeDescriptor(type);
-    }
-
-    private void assertArray(Object actual, int index) {
-        switch (argTypes[index]) {
-            case 9:
-                Assert.assertArrayEquals((byte[]) argTuples[index].value, (byte[]) actual);
-                break;
-            case 10:
-                Assert.assertArrayEquals((boolean[]) argTuples[index].value, (boolean[]) actual);
-                break;
-            case 11:
-                Assert.assertArrayEquals((char[]) argTuples[index].value, (char[]) actual);
-                break;
-            case 12:
-                Assert.assertArrayEquals((short[]) argTuples[index].value, (short[]) actual);
-                break;
-            case 13:
-                Assert.assertArrayEquals((int[]) argTuples[index].value, (int[]) actual);
-                break;
-            case 14:
-                Assert.assertArrayEquals((long[]) argTuples[index].value, (long[]) actual);
-                break;
-            case 15:
-                Assert.assertTrue(Arrays.equals((float[]) argTuples[index].value, (float[]) actual));
-                break;
-            case 16:
-                Assert.assertArrayEquals((double[]) argTuples[index].value, (double[]) actual, 0.1);
-                break;
-            case 18:
-                Assert.assertArrayEquals((int[][]) argTuples[index].value, (int[][]) actual);
-                break;
-            default:
-                Assert.fail("Not an array, test failed");
-        }
     }
 }
